Fix: melhorando métodos do EstoqueService

+++ src/main/java/com/trier/futmax/repository/ProdutoRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ProdutoRepository extends JpaRepository<ProdutoModel, Integer> {
+public interface ProdutoRepository extends JpaRepository<ProdutoModel, Long> {
 
     Optional<ProdutoModel> findByCdProduto(Long cdProduto);
